first test for this low level part of this package:
It shows there are two functions, living (currently) in
a module called `event_io_file`, since they kind of replace
what was an EventIOFile before.

One returns a (nested) list of only ObjectHeaders, which
can serve as a starting point for other code to find the
payload inside the file, do something with the payload,
and give higher level stuff to a possible user.

The other function yields tuples of ([a list of ObjectHeaders], io.Bytes)