fix test_n_events:
IACTFile has not __len__. We have to discuss what we want __len__ to be:
 - n_showers?
 - n_showers * n_reuse?
 - ???

For example, can it e, that for a certain reuse, there are no photons at the
detector, so that the total number of "events" is smaller than
n_showers * n_reuse? This was the case in old MMCS times.