make payload_has_correct_size test pass
problem was, that for sub-level objects, the `file` is not the
actual file, but a io.BytesIO object. So the start_adress should not
be counted from the start of the file, but from the start if the
BytesIO object, which is given by the last header in the list
of headers again.

However, this increases the memory footprint, since the
`file` object is not really a file, but a BytesIO, which
can be large.

Add a test, to make sure `EventIOObject._file` is really a file